+++ src/main/java/de/sakurajin/evenbetterarcheology/api/ArtifactEnchantment.java
@@ -4,7 +4,7 @@ import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentTarget;
 import net.minecraft.entity.EquipmentSlot;
 
-public class ArtifactEnchantment extends Enchantment {
+public abstract class ArtifactEnchantment extends Enchantment {
     protected ArtifactEnchantment(Rarity weight, EnchantmentTarget target, EquipmentSlot... slotTypes) {
         super(weight, target, slotTypes);
     }