+++ src/main/java/de/sakurajin/evenbetterarcheology/enchantment/ModEnchantments.java
@@ -1,12 +1,14 @@
 package de.sakurajin.evenbetterarcheology.enchantment;
 
 import de.sakurajin.evenbetterarcheology.EvenBetterArcheology;
+import de.sakurajin.evenbetterarcheology.api.ArtifactEnchantment;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
+import net.fabricmc.fabric.api.entity.event.v1.FabricElytraItem;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentLevelEntry;
 import net.minecraft.enchantment.EnchantmentTarget;
 import net.minecraft.entity.EquipmentSlot;
-import net.minecraft.item.EnchantedBookItem;
+import net.minecraft.item.*;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.text.Text;
@@ -15,8 +17,31 @@ import net.minecraft.util.Identifier;
 
 public class ModEnchantments {
     public static Enchantment PENETRATING_STRIKE = register("penetrating_strike", new PenetratingStrikeEnchantment(Enchantment.Rarity.VERY_RARE, EquipmentSlot.MAINHAND));
-    public static Enchantment SOARING_WINDS = register("soaring_winds", new SoaringWindsEnchantment(Enchantment.Rarity.VERY_RARE, EnchantmentTarget.ARMOR_CHEST, EquipmentSlot.MAINHAND));
-    public static Enchantment TUNNELING = register("tunneling", new TunnelingEnchantment(Enchantment.Rarity.VERY_RARE, EnchantmentTarget.DIGGER, EquipmentSlot.MAINHAND) {
+    public static Enchantment SOARING_WINDS = register("soaring_winds", new ArtifactEnchantment(Enchantment.Rarity.VERY_RARE, EnchantmentTarget.ARMOR_CHEST, EquipmentSlot.MAINHAND) {
+        @Override
+        public int getMaxLevel() {
+            return EvenBetterArcheology.CONFIG.SOARING_WINDS_MAXLEVEL();
+        }
+
+        @Override
+        public boolean isAcceptableItem(ItemStack stack) {
+
+            return stack.getItem() instanceof ElytraItem || stack.getItem() instanceof FabricElytraItem;
+        }
+    });
+
+    public static Enchantment TUNNELING = register("tunneling", new ArtifactEnchantment(Enchantment.Rarity.VERY_RARE, EnchantmentTarget.DIGGER, EquipmentSlot.MAINHAND){
+        @Override
+        public boolean isAcceptableItem(ItemStack stack) {
+            return stack.getItem() instanceof PickaxeItem || stack.getItem() instanceof ShovelItem;
+        }
+    });
+
+    public static Enchantment SEAS_BOUNTY = register("seas_bounty", new ArtifactEnchantment(Enchantment.Rarity.VERY_RARE, EnchantmentTarget.FISHING_ROD, EquipmentSlot.MAINHAND) {
+        @Override
+        public boolean isAcceptableItem(ItemStack stack) {
+            return stack.getItem() instanceof FishingRodItem;
+        }
     });
 
     private static Enchantment register(String name, Enchantment enchantment) {